core: add pandoc bibliography
Use a post processing script to replace [@key] with the corresponding
citation, and append at the end of the post a list of bibliography.

The script run the original file, which includes the full bibliography
in CSL YAML format in its front matter, through pandoc with `--citeproc`
enabled, and use the resulting HTML to collect those citations and the
bibliography.